Move Role to project-member. Remove team-member and user-role.
