+++ src/components/project-member/dto/index.ts
@@ -1,3 +1,4 @@
+export * from './role.dto';
 export * from './project-member.dto';
 export * from './create-project-member.dto';
 export * from './update-project-member.dto';
