+++ src/components/project-member/dto/update-project-member.dto.ts
@@ -1,8 +1,8 @@
 import { Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 import { Field, ID, InputType, ObjectType } from 'type-graphql';
-import { Role } from '../../user/role';
 import { ProjectMember } from './project-member.dto';
+import { Role } from './role.dto';
 
 @InputType()
 export abstract class UpdateProjectMember {
