--- src/components/user/user-role.ts
@@ -1,12 +1,0 @@
-import { InputType, ObjectType } from 'type-graphql';
-import { Location } from '../location';
-import { Role } from './role';
-
-export interface UserRole {
-  role: Role;
-  locations: Location[];
-}
-
-@ObjectType()
-@InputType('UserRoleInput')
-export class UserRole {}
